fetch-service/berlin_intensivecare/requirements.txt
pycouchdb
python-dotenv
geojson
urllib3
